delete item permission done

+++ backend/src/resolvers/Mutation.js
@@ -41,9 +41,13 @@ const Mutations = {
   async deleteItem(parent, args, ctx, info) {
     const where = { id: args.id }
     //find item
-    const item = await ctx.db.query.item({ where }, `{id title}`)
+    const item = await ctx.db.query.item({ where }, `{id title user { id }}`)
     //check if user has permission
-    //TODO
+    const ownsItem = item.user.id === ctx.request.userId
+    const hasPermission = ctx.request.user.permissions.some(permission => ['ADMIN', 'ITEMDELETE'].includes(permission))
+    if (!ownsItem && !hasPermission) {
+      throw new Error('You do not have permission to delete this item!')
+    }
     //delete item
     return ctx.db.mutation.deleteItem({ where }, info)
   },
